% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4preProcess.R
\name{calc_Chl}
\alias{calc_Chl}
\title{wet Chl calc}
\usage{
calc_Chl(df)
}
\arguments{
\item{df}{data.frame (with A665, A649, A470, vol, area and mass column)}
}
\value{
modified df
}
\description{
wet Chl calc
}
