% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{collect_caret_metrics_fullStage}
\alias{collect_caret_metrics_fullStage}
\title{collect gof from do_caret_VI and do_caret_bands result(full stages fit)}
\usage{
collect_caret_metrics_fullStage(fit)
}
\arguments{
\item{fit}{the do_caret_VI and do_caret_bands result}
}
\value{
tibble
}
\description{
collect gof from do_caret_VI and do_caret_bands result(full stages fit)
}
