% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{collect_caret_ObsVsPred}
\alias{collect_caret_ObsVsPred}
\title{obs vs pred plot from do_caret_VI and do_caret_bands result}
\usage{
collect_caret_ObsVsPred(
  fit,
  RMSE_precision = 2,
  ylabel = NULL,
  isNormalzed = FALSE,
  isCV = FALSE
)
}
\arguments{
\item{fit}{the do_caret_VI and do_caret_bands result}

\item{RMSE_precision}{n digits for RMSE annotate}

\item{ylabel}{target variable label}

\item{isNormalzed}{control wether the RMSE or nRMSE annotate}

\item{isCV}{control wether the cv or train gof annotate}
}
\value{

}
\description{
obs vs pred plot from do_caret_VI and do_caret_bands result
}
