% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{collect_caret_VIcurve}
\alias{collect_caret_VIcurve}
\title{generate df_curve from do_caret_VI result}
\usage{
collect_caret_VIcurve(fit)
}
\arguments{
\item{fit}{the do_caret_VI fit}
}
\value{
df
}
\description{
generate df_curve from do_caret_VI result
}
