% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{createDataPartition_2col}
\alias{createDataPartition_2col}
\title{inTrain/Test split wrapper -> add inTrain column for a df}
\usage{
createDataPartition_2col(
  df,
  yname,
  how = "random",
  p = 0.75,
  groups = 5,
  step = 4
)
}
\arguments{
\item{df}{the target df}

\item{yname}{the column name will used for stratification}

\item{p}{the percentage used for standalone validation}

\item{groups}{n groups for stratification}
}
\value{
df with addictional inTrain column
}
\description{
inTrain/Test split wrapper -> add inTrain column for a df
}
