% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{createDataPartition_2PlotIDdict}
\alias{createDataPartition_2PlotIDdict}
\title{inTrain/Test split wrapper -> return list of inTrain/Test PlotID}
\usage{
createDataPartition_2PlotIDdict(df, yname, p = 0.66, groups = 5)
}
\arguments{
\item{df}{the target df}

\item{yname}{the column name will used for stratification}

\item{p}{the percentaget used for standalone validation}

\item{groups}{n groups for stratification}
}
\value{
list of vector of PlotID fall in inTrain/Test
}
\description{
inTrain/Test split wrapper -> return list of inTrain/Test PlotID
}
