% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{do_caret_bands}
\alias{do_caret_bands}
\title{caret::train wrapper to fit biochemphy~bands machine learning models}
\usage{
do_caret_bands(
  method,
  tuneGrid,
  tuneLength,
  fm,
  df_inTrain,
  df_Test,
  indexCV,
  ylabel = NULL,
  preProcess_fun = c("center", "scale")
)
}
\arguments{
\item{method}{the caret::train method param}

\item{tuneGrid}{the caret::trainControl tuneGrid param}

\item{tuneLength}{the caret::trainControl tuneLength param}

\item{fm}{the caret::train fm param}

\item{df_inTrain}{df inTrain}

\item{df_Test}{df for standalone validation}

\item{indexCV}{cross-validation split of df_inTrain}

\item{ylabel}{label for target variable}
}
\value{

}
\description{
caret::train wrapper to fit biochemphy~bands machine learning models
}
