% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{do_caret_VI}
\alias{do_caret_VI}
\title{caret::train wrapper to fit biochemphy~index curve}
\usage{
do_caret_VI(index, spc, biochemphy, fun_indexCV = get_indexCV)
}
\arguments{
\item{index}{the index name for ssdxj_vegindex}

\item{spc}{the speclib obj}

\item{biochemphy}{the target variable name in SI}
}
\value{
a modified caret::obj. With x(index), y(biochemphy), obs, pred in df_inTrain and df_Test
}
\description{
caret::train wrapper to fit biochemphy~index curve
}
