% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{doAddinTrain}
\alias{doAddinTrain}
\title{inTrain/Test split wrapper}
\usage{
doAddinTrain(df, yname, p = 0.66, groups = 5)
}
\arguments{
\item{groups}{}
}
\value{
df with inTrain column
}
\description{
inTrain/Test split wrapper
}
