% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{doAddinTrain_PlotID}
\alias{doAddinTrain_PlotID}
\title{inTrain/Test split wrapper}
\usage{
doAddinTrain_PlotID(df, yname, p = 0.66, groups = 5)
}
\arguments{
\item{groups}{}
}
\value{
vector of PlotIDs fall in inTrain/Test
}
\description{
inTrain/Test split wrapper
}
