% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVI.R
\name{.calc_mREIP}
\alias{.calc_mREIP}
\title{helper to get REP using Gaussain fit, because the \code{hsdar::vegindex(spc, 'mREIP')} DO NOT add sigma}
\usage{
.calc_mREIP(spc)
}
\arguments{
\item{spc}{the speclib obj}
}
\value{
numerical vector
}
\description{
helper to get REP using Gaussain fit, because the \code{hsdar::vegindex(spc, 'mREIP')} DO NOT add sigma
}
