% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVI.R
\name{.calc_NAOC}
\alias{.calc_NAOC}
\title{helper for VI: NAOC}
\usage{
.calc_NAOC(spc, a = 643, b = 795)
}
\arguments{
\item{spc}{the speclib obj}

\item{a}{start band (in nm)}

\item{b}{end band (in nm)}
}
\value{
numerical vector
}
\description{
helper for VI: NAOC
}
