% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVI.R
\name{.calc_REP_multi}
\alias{.calc_REP_multi}
\title{helper to get REP considering multi REP simuation}
\usage{
.calc_REP_multi(spc)
}
\arguments{
\item{spc}{}
}
\value{
a data.frame
}
\description{
helper to get REP considering multi REP simuation
}
