% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVI.R
\name{.calc_REP_sg}
\alias{.calc_REP_sg}
\title{helper to get REP by defination (max d1)}
\usage{
.calc_REP_sg(spc)
}
\arguments{
\item{spc}{the speclib obj}
}
\value{
numerical vector
}
\description{
helper to get REP by defination (max d1)
}
