% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{.print_error}
\alias{.print_error}
\title{helper for \code{stop()} function}
\usage{
.print_error(e, msg)
}
\arguments{
\item{e}{error obj}

\item{msg}{additional msg}
}
\description{
helper for \code{stop()} function
}
