% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{get_indexCV}
\alias{get_indexCV}
\title{Create repeated n fold corss-validation index}
\usage{
get_indexCV(x, folds = 10, times = 5)
}
\arguments{
\item{x}{numerical vector(normally the inTrain part of response vector)}

\item{times}{n times}

\item{fold}{n folds}
}
\value{
list of cv index
}
\description{
Create repeated n fold corss-validation index
}
