% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{get_inTrain_fromTestSPC}
\alias{get_inTrain_fromTestSPC}
\title{Title get inTrain index by full dataset and Test dataset}
\usage{
get_inTrain_fromTestSPC(input_full, input_Test)
}
\arguments{
\item{input_full}{full spc (or SI)}

\item{input_Test}{test spc (or SI)}
}
\value{
inTrain index vector
}
\description{
Title get inTrain index by full dataset and Test dataset
}
