% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{gglm}
\alias{gglm}
\title{shortcut obs vs pred scatter plot for 'lm' obj}
\usage{
gglm(fit, smooth = "lm")
}
\arguments{
\item{fit}{'lm' obj}

\item{smooth}{stat_smooth method}
}
\value{
p(ggplot2)
}
\description{
shortcut obs vs pred scatter plot for 'lm' obj
}
\examples{
gglm(lm(y~x))
}
