% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4keras.R
\name{layer_DenseTransposeTiedLayer}
\alias{layer_DenseTransposeTiedLayer}
\title{warpper to generate keras tied layer}
\usage{
layer_DenseTransposeTiedLayer(object, output_dim, tied_to)
}
\arguments{
\item{object}{tensor obj}

\item{output_dim}{output dim}

\item{tied_to}{layer tie to}
}
\value{
layer
}
\description{
warpper to generate keras tied layer
}
