% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_LbyL.R
\name{LbyL_resp_fun_SR}
\alias{LbyL_resp_fun_SR}
\title{resp_fun for \code{\link{wrapper_LbyL}} to calc SR index}
\usage{
LbyL_resp_fun_SR(b1, b2)
}
\description{
resp_fun for \code{\link{wrapper_LbyL}} to calc SR index
}
