% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_LbyL.R
\name{LbyL_stat_fun_r2}
\alias{LbyL_stat_fun_r2}
\title{stat_fun for \code{\link{wrapper_LbyL}} to calc r2 of lm fit}
\usage{
LbyL_stat_fun_r2(x, y)
}
\description{
stat_fun for \code{\link{wrapper_LbyL}} to calc r2 of lm fit
}
