% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfos.R
\name{modelInfo_bear}
\alias{modelInfo_bear}
\title{modelInfo for caret doing bear's law VI curve fitting}
\usage{
modelInfo_bear()
}
\value{
the modelInfo(list)
}
\description{
modelInfo for caret doing bear's law VI curve fitting
}
