% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfo_bearReverse.R
\name{modelInfo_bearReverse}
\alias{modelInfo_bearReverse}
\title{modelInfo for caret doing bear's law VI curve fitting}
\usage{
modelInfo_bearReverse()
}
\value{
the modelInfo(list)
}
\description{
modelInfo for caret doing bear's law VI curve fitting
}
