% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfo_exp.R
\name{modelInfo_exp}
\alias{modelInfo_exp}
\title{modelInfo for caret doing exp VI curve fitting}
\usage{
modelInfo_exp()
}
\value{
the modelInfo(list)
}
\description{
modelInfo for caret doing exp VI curve fitting
}
