% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInfos.R
\name{modelInfo_rf_mtry_ntree}
\alias{modelInfo_rf_mtry_ntree}
\title{modelInfo fore caret doing rf tuning both mtry and ntree}
\usage{
modelInfo_rf_mtry_ntree()
}
\value{
the modelInfo(list)
}
\description{
modelInfo fore caret doing rf tuning both mtry and ntree
}
