% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parser_cubert_export}
\alias{parser_cubert_export}
\title{parser for Cubert export txt file}
\usage{
parser_cubert_export(fpath)
}
\arguments{
\item{fpath}{}
}
\value{

}
\description{
parser for Cubert export txt file
}
