% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parser_ENVIROIstat}
\alias{parser_ENVIROIstat}
\title{parser for ENVI ROI stat file, the ROIs are from shpfile by "unique records of an attribute to seperate ROIs" with 'PlotID" attribute}
\usage{
parser_ENVIROIstat(fpath, gain = 1e-04)
}
\arguments{
\item{fpath}{the txt file path}

\item{gain}{dafualt 0.0001 for s185}
}
\value{
df
}
\description{
parser for ENVI ROI stat file, the ROIs are from shpfile by "unique records of an attribute to seperate ROIs" with 'PlotID" attribute
}
