% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parser_sig}
\alias{parser_sig}
\title{parser of single HR1024i sig file}
\usage{
parser_sig(fpath)
}
\arguments{
\item{fpath:}{fpath of sig file. The Field, SampleDate, isCanopy, PlotID,
SampleID and RepID are extracted from the fname.
So fname must in format 'YNSD_20140712_C_01_01_00_moc_resamp.sig'
which is 'FieldID_SampleDate_isCanopy_PlotID_SampleID_RepID_others.sig'}
}
\value{
parsed content(list)
}
\description{
parser of single HR1024i sig file
}
