% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parser_sig_2df}
\alias{parser_sig_2df}
\title{a svc sig file parser}
\usage{
parser_sig_2df(fpath)
}
\arguments{
\item{fpath}{the svc sig file path}
}
\value{
a long tibble
}
\description{
a svc sig file parser
}
