% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parser_sigs}
\alias{parser_sigs}
\title{Wrapper of \code{\link{parser_sig}}, to parse all sig files in a dir.}
\usage{
parser_sigs(dir)
}
\arguments{
\item{dir:}{dir contains sig files}
}
\value{
list of parsed content(list)
}
\description{
Wrapper of \code{\link{parser_sig}}, to parse all sig files in a dir.
}
\references{
parser_sig()
}
