% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parser_sunScan}
\alias{parser_sunScan}
\title{parser of SunScan file}
\usage{
parser_sunScan(fpath)
}
\arguments{
\item{fpath}{fpath of SunScan file}
}
\value{
parsed df
}
\description{
parser of SunScan file
}
