% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4keras.R
\name{predict_keras_spc}
\alias{predict_keras_spc}
\title{wrapper for do predict of SAE models and turn result to spc}
\usage{
predict_keras_spc(model, spc, wl = NULL)
}
\arguments{
\item{model}{keras(SAE) model}

\item{spc}{input spc obj}

\item{wl}{\describe{
 \item{NULL}{for encoder model, use 1:n channel for wavelength}
 \item{nmerical vector}{for decoder model}
}}
}
\value{
spc obj
}
\description{
wrapper for do predict of SAE models and turn result to spc
}
