% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4preProcess.R
\name{prepare_obj4wf}
\alias{prepare_obj4wf}
\title{prepare obj (list(inTrain, spc_full, indexCV)) for training workflow
\describe{
\item{inTrain}{\code{\link[G407]{get_inTrain_respOrder_withGroup}}}
\item{spc_full}{the spc}
\item{indexCV}{\code{\link[G407]{get_indexCV}}}
}}
\usage{
prepare_obj4wf(spc, biochemphy, group, folds, times, isSplit)
}
\arguments{
\item{spc}{ful spc}

\item{biochemphy}{name of response in SI}

\item{group}{name of group in SI}

\item{folds}{for cv}

\item{times}{for cv}

\item{isSplit}{whether inTrain/Test split}
}
\value{
list(inTrain, spc_full, indexCV)
}
\description{
prepare obj (list(inTrain, spc_full, indexCV)) for training workflow
\describe{
\item{inTrain}{\code{\link[G407]{get_inTrain_respOrder_withGroup}}}
\item{spc_full}{the spc}
\item{indexCV}{\code{\link[G407]{get_indexCV}}}
}
}
