% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{pretty_value}
\alias{pretty_value}
\title{format digits in training param to str}
\usage{
pretty_value(x)
}
\arguments{
\item{x}{input value}
}
\value{
formated value
}
\description{
format digits in training param to str
}
