% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{pretty_values}
\alias{pretty_values}
\title{format digits in training param to str}
\usage{
pretty_values(x)
}
\arguments{
\item{x}{input value vector}
}
\value{
formated value vector
}
\description{
format digits in training param to str
}
