% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4PRO4SAIL_D.R
\name{PRO4SAIL_D}
\alias{PRO4SAIL_D}
\title{modeling reflectance data from canopy and  modeling leaf optical properties with
PROSPECT-D (Feret et al. 2017)}
\usage{
PRO4SAIL_D(
  N,
  Cab,
  Car,
  Cant,
  Cbrown,
  Cw,
  Cm,
  LIDFa,
  LIDFb,
  TypeLidf,
  LAI,
  hspot,
  tts,
  tto,
  psi,
  rsoil
)
}
\arguments{
\item{rsoil}{}
}
\value{

}
\description{
modeling reflectance data from canopy and  modeling leaf optical properties with
PROSPECT-D (Feret et al. 2017)
}
