% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4PRO4SAIL_D.R
\name{PRO4SAIL_D_wrapper_df}
\alias{PRO4SAIL_D_wrapper_df}
\title{wrapper for PRO4SAIL_D with parameter_list input, handling skyl}
\usage{
PRO4SAIL_D_wrapper_df(df, rsoil = NULL)
}
\arguments{
\item{rsoil}{}
}
\value{

}
\description{
wrapper for PRO4SAIL_D with parameter_list input, handling skyl
}
