% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4PRO4SAIL_D.R
\name{PRO4SAIL_DB_wrapper_var}
\alias{PRO4SAIL_DB_wrapper_var}
\title{wrapper for PRO4SAIL_D, handling skyl}
\usage{
PRO4SAIL_DB_wrapper_var(
  Cab,
  Car,
  Cant,
  Cbrown,
  Cw,
  Cm,
  N,
  psoil,
  LAI,
  hspot,
  tts,
  tto,
  psi,
  TypeLidf,
  LIDFa,
  LIDFb,
  rsoil = NULL
)
}
\arguments{
\item{rsoil}{}
}
\value{

}
\description{
wrapper for PRO4SAIL_D, handling skyl
}
