% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4PRO4SAIL_D.R
\name{PROSPECT_D}
\alias{PROSPECT_D}
\title{Modeling leaf optical properties with PROSPECT-D (Feret et al. 2017)}
\usage{
PROSPECT_D(N, Cab, Car, Cant, Cbrown, Cw, Cm)
}
\arguments{
\item{Cm}{}
}
\value{

}
\description{
Modeling leaf optical properties with PROSPECT-D (Feret et al. 2017)
}
