% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4tidymodels.R
\name{spc_2bandVI_test}
\alias{spc_2bandVI_test}
\title{A quick check of a new  two band index by give bands and function}
\usage{
spc_2bandVI_test(spc, wl1, wl2, fun = function(r1, r2) {
    r1/r2
})
}
\arguments{
\item{spc}{the speclib obj}

\item{wl1}{band1}

\item{wl2}{band2}

\item{fun}{define of index}
}
\value{

}
\description{
A quick check of a new  two band index by give bands and function
}
