% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_2df}
\alias{spc_2df}
\title{Transform Speclib to data.frame (spectral columns name is in '450' format)}
\usage{
spc_2df(spc)
}
\arguments{
\item{spc:}{'speclib' obj}
}
\value{
'data.frame' obj
}
\description{
Transform Speclib to data.frame (spectral columns name is in '450' format)
}
