% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_2df_melt}
\alias{spc_2df_melt}
\title{spc or spc_df to long df}
\usage{
spc_2df_melt(input)
}
\arguments{
\item{input}{speclib obj or df}
}
\value{
df (long df with "wl" and "reflect" column)
}
\description{
spc or spc_df to long df
}
