% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_2df_meltWithMask}
\alias{spc_2df_meltWithMask}
\title{spc to long df with masked bands as NA for ggplot}
\usage{
spc_2df_meltWithMask(spc)
}
\arguments{
\item{spc}{}
}
\value{
df (long df with "wl" and "reflect" column)
}
\description{
spc to long df with masked bands as NA for ggplot
}
