% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_2dfB}
\alias{spc_2dfB}
\title{transform Speclib into data.frame (spectral columns name is in 'B_450' format)}
\usage{
spc_2dfB(spc)
}
\arguments{
\item{spc:}{'speclib' obj}
}
\value{
'data.frame' obj
}
\description{
transform Speclib into data.frame (spectral columns name is in 'B_450' format)
}
