% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_addNoise}
\alias{spc_addNoise}
\title{add gaussian noise to a speclib}
\usage{
spc_addNoise(spc_sim, noise_base, noise_level = 0.04)
}
\arguments{
\item{noise_level}{}
}
\value{

}
\description{
add gaussian noise to a speclib
}
