% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_bandwiseCor}
\alias{spc_bandwiseCor}
\title{spc bandwise cor}
\usage{
spc_bandwiseCor(spc, yname, colLabel)
}
\arguments{
\item{spc}{the speclib object}

\item{yname}{meta colname}
}
\value{
df data.frame with colnames c(wl, colLabel)
}
\description{
spc bandwise cor
}
