% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_cor}
\alias{spc_cor}
\title{Bandwise cor with responser}
\usage{
spc_cor(spc, biochemphy)
}
\arguments{
\item{spc}{spc}

\item{biochemphy}{name of responsor in SI}
}
\value{
df with col ('wl', 'estimate', 'p.value')
}
\description{
Bandwise cor with responser
}
