% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_cor_stage}
\alias{spc_cor_stage}
\title{wrapper of \code{\link{spc_cor}} doing bandwise band cor with responser
grouped by stage}
\usage{
spc_cor_stage(stageValue, spc, biochemphy)
}
\arguments{
\item{stageValue}{levels of stage in SI}

\item{spc}{spc}

\item{biochemphy}{name of reponser in SI}
}
\value{
df
}
\description{
wrapper of \code{\link{spc_cor}} doing bandwise band cor with responser
grouped by stage
}
