% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{spc_from_sigs}
\alias{spc_from_sigs}
\title{generate spc obj form parsed contents}
\usage{
spc_from_sigs(dat)
}
\arguments{
\item{dat}{result from parser_sigs}
}
\value{
df df of parsed result
}
\description{
generate spc obj form parsed contents
}
