% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_fromDf}
\alias{spc_fromDf}
\title{generate 'Speclib' obj from data.frame, reflectance selected by matches
colnames matches('^(\\d)+(\\.\\d+)?$'),wl determined by colnames of spectral data}
\usage{
spc_fromDf(df, bands_reg = "^(\\\\d)+(\\\\.\\\\d+)?$")
}
\arguments{
\item{df}{data.frame of data}

\item{bands_reg}{with default '^(\\d)+(\\.\\d+)?$'}
}
\value{
Speclib obj
}
\description{
generate 'Speclib' obj from data.frame, reflectance selected by matches
colnames matches('^(\\d)+(\\.\\d+)?$'),wl determined by colnames of spectral data
}
