% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_generator}
\alias{spc_generator}
\title{genearate Speclib obj from: csv file or data.frame}
\usage{
spc_generator(input, idCol = "PlotID")
}
\arguments{
\item{input}{csv file or data.frame}

\item{idCol}{colname of id col ("PlotID")}
}
\value{
Speclib obj
}
\description{
genearate Speclib obj from: csv file or data.frame
}
