% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_greenPeak}
\alias{spc_greenPeak}
\title{get green peak position (vec) from give spc by finding the max reflect between 500 and 660nm}
\usage{
spc_greenPeak(spc)
}
\arguments{
\item{spc}{the spc dataset}
}
\value{
green peak position vector
}
\description{
get green peak position (vec) from give spc by finding the max reflect between 500 and 660nm
}
