% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4tidymodels.R
\name{spc_index_scatter}
\alias{spc_index_scatter}
\title{scatter plot between biochemphy and index from spc and indexName}
\usage{
spc_index_scatter(spc, indexName, biochemphy)
}
\arguments{
\item{spc}{the speclib obj}

\item{indexName}{the name of index}

\item{biochemphy}{the name of biochemphy}
}
\value{
a ggplot obj
}
\description{
scatter plot between biochemphy and index from spc and indexName
}
