% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_inTrain}
\alias{spc_inTrain}
\title{get inTrain spc by index vector}
\usage{
spc_inTrain(spc, inTrain)
}
\arguments{
\item{spc:}{spc to be filtered}

\item{inTrain:}{inTrain index vector}
}
\value{
spc obj
}
\description{
get inTrain spc by index vector
}
