% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_melt}
\alias{spc_melt}
\title{melt the reflectance part of spc df}
\usage{
spc_melt(x, band_reg = "^(\\\\d)+(\\\\.\\\\d+)?$")
}
\arguments{
\item{x}{spc or spc_df}

\item{band_reg}{with default '^(\\d)+(\\.\\d+)?$'}
}
\value{
tibble
}
\description{
melt the reflectance part of spc df
}
