% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R, R/ulits4spc.R
\name{spc_plot}
\alias{spc_plot}
\title{plot speclib}
\usage{
spc_plot(spc, mask = NA)

spc_plot(spc, mask = NA)
}
\arguments{
\item{spc}{hsdar::speclib obj}

\item{mask}{NA or vector}

\item{x}{spc obj}

\item{FUN}{ids of spectral(vector) or NULL}
}
\value{
ggplot

ggplot
}
\description{
plot speclib

plot hsdar::speclib obj (one line per on record)
}
