% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{spc_plot2}
\alias{spc_plot2}
\title{Speclib plot(all or selected PlotID)}
\usage{
spc_plot2(spc, idCol = "PlotID", idFilter = NULL, ylims = c(0, 0.8))
}
\arguments{
\item{spc}{the Speclib object}

\item{idCol}{characters, the colname of groups (ex. PlotID)}

\item{idFilter}{vector of characters(c("P01", "P02"))}

\item{ylims}{}
}
\value{
ggplot obj
}
\description{
Speclib plot(all or selected PlotID)
}
