% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{spc_plot_all}
\alias{spc_plot_all}
\title{plot all spectra of a speclib}
\usage{
spc_plot_all(spc, alpha = 1/5)
}
\arguments{
\item{spc}{the speclib obj}

\item{alpha}{for geom_line}
}
\value{
a ggplot obj
}
\description{
plot all spectra of a speclib
}
