% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{spc_plot_filters}
\alias{spc_plot_filters}
\title{plot selected spectra of a speclib, selected by index vector or a data.frame}
\usage{
spc_plot_filters(spc, by, alpha = 1)
}
\arguments{
\item{spc}{speclib obj}

\item{by}{a (numeric) index vector or data.frame for left_join}

\item{alpha}{for geom_line}
}
\value{
a ggplot obj
}
\description{
plot selected spectra of a speclib, selected by index vector or a data.frame
}
