% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{spc_plot_sd}
\alias{spc_plot_sd}
\title{plot mean, mean+sd and mean-sd spectra of ad speclib}
\usage{
spc_plot_sd(spc)
}
\arguments{
\item{spc}{the speclib obj}
}
\value{
a ggplot obj
}
\description{
plot mean, mean+sd and mean-sd spectra of ad speclib
}
