% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_pop}
\alias{spc_pop}
\title{spc reflectance to list of one spectra spc}
\usage{
spc_pop(spc)
}
\arguments{
\item{spc}{the spclib obj}

\item{tag}{for name}
}
\value{
list of vectors
}
\description{
spc reflectance to list of one spectra spc
}
