% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_resample}
\alias{spc_resample}
\title{resample spc_sim tospc_obs spectra resolution}
\usage{
spc_resample(spc_sim, spc_obs)
}
\arguments{
\item{spc_obs}{}
}
\value{

}
\description{
resample spc_sim tospc_obs spectra resolution
}
