% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_rsoil_bs}
\alias{spc_rsoil_bs}
\title{do spc * bs}
\usage{
spc_rsoil_bs(rsoil, bs = seq(0.5, 1.5, length.out = 7))
}
\arguments{
\item{rsoil}{a one spectra spc}

\item{bs}{the bs vector}
}
\value{
a 1*length(bs) spectra spc
}
\description{
do spc * bs
}
