% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_rsoil_s185_expand}
\alias{spc_rsoil_s185_expand}
\title{expand s185 spectra(rsoil) to 400:2500 nm by fwhm 1nm (for PROSAIL)}
\usage{
spc_rsoil_s185_expand(spc, ref_default = 1e-04)
}
\arguments{
\item{spc}{a one spectra speclib obj}

\item{ref_default}{reflectance of bands beyound 470:950 will be set to ref_default}
}
\value{
a speclib obj
}
\description{
expand s185 spectra(rsoil) to 400:2500 nm by fwhm 1nm (for PROSAIL)
}
