% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4tidymodels.R
\name{spc_test_inTrain}
\alias{spc_test_inTrain}
\title{a boxplot for check the inTrain/Test split}
\usage{
spc_test_inTrain(spc, biochemphy)
}
\arguments{
\item{spc}{the speclib}

\item{biochemphy}{the target var}
}
\value{

}
\description{
a boxplot for check the inTrain/Test split
}
