% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spcdf_mask2NA}
\alias{spcdf_mask2NA}
\title{change the reflectance of masked bands to NA}
\usage{
spcdf_mask2NA(input, mask = NA)
}
\arguments{
\item{input}{spc or melt spc_df}

\item{mask}{NA or vector}
}
\value{
df
}
\description{
change the reflectance of masked bands to NA
}
