% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVI.R
\name{ssdxj_vegindex}
\alias{ssdxj_vegindex}
\title{lapply wrapper for hsdar::vegetation index}
\usage{
ssdxj_vegindex(index, spc, weighted = FALSE, ...)
}
\arguments{
\item{index:}{name of index}

\item{spc:}{the speclib obj}
}
\value{
numeric vector of VI values
}
\description{
lapply wrapper for hsdar::vegetation index
}
