% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{themeDotplot}
\alias{themeDotplot}
\title{theme for dotchart plot}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
themeDotplot
}
\value{
ggplot theme
}
\description{
theme for dotchart plot
}
\keyword{datasets}
