% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4tidymodels.R
\name{tidymodels_linearReg}
\alias{tidymodels_linearReg}
\title{tidymodel routine for fitting a linear VI vs biochemphy relationship}
\usage{
tidymodels_linearReg(spc, indexName, biochemphy)
}
\arguments{
\item{spc}{the speclib obj}

\item{indexName}{the index name}

\item{biochemphy}{the biochemphy name}
}
\value{
a tidymodel obj
}
\description{
tidymodel routine for fitting a linear VI vs biochemphy relationship
}
