% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4tidymodels.R
\name{tidymodels_ML}
\alias{tidymodels_ML}
\title{tidymodel routine for ML fitting}
\usage{
tidymodels_ML(spec, res, df_training, df_testing, df_folds)
}
\arguments{
\item{spec}{model_spec obj}

\item{res}{(data) recipe obj}

\item{df_training}{df for training}

\item{df_testing}{df for stand-alone validation}

\item{df_folds}{info for cross-validation}
}
\value{
list(cv, df_training, df_testing, df_metrics)
}
\description{
tidymodel routine for ML fitting
}
