% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{train_add_coefs}
\alias{train_add_coefs}
\title{get bear coefs from caret::train obj}
\usage{
train_add_coefs(trainObj)
}
\arguments{
\item{trainObj}{}
}
\value{
charactor
}
\description{
get bear coefs from caret::train obj
}
