% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{train_add_curveDf}
\alias{train_add_curveDf}
\title{update caret::train obj with add_curveDf, to simulate the trained function (curve)
where x is \code{x <- seq(min(df$vi), max(df$vi), length.out = 1000)}, while y is
calculated by parse the coeffciences.
handled method include: \itemize{
  \item method == 'lm'
  \item method == 'lm2'
  \item method == 'bear'
}}
\usage{
train_add_curveDf(train_obj)
}
\arguments{
\item{trian_obj:}{caret::train obj}
}
\value{
df
}
\description{
update caret::train obj with add_curveDf, to simulate the trained function (curve)
where x is \code{x <- seq(min(df$vi), max(df$vi), length.out = 1000)}, while y is
calculated by parse the coeffciences.
handled method include: \itemize{
  \item method == 'lm'
  \item method == 'lm2'
  \item method == 'bear'
}
}
