% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{train_add_gof}
\alias{train_add_gof}
\title{get train and test gof from fit obj (depend add_trainDf and add_testDf in fit obj)}
\usage{
train_add_gof(train_obj)
}
\arguments{
\item{train_obj}{caret::train (caret::rfe) obj
(updated version with add_trainDf and add_testDf)}
}
\value{
df
}
\description{
get train and test gof from fit obj (depend add_trainDf and add_testDf in fit obj)
}
