% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{train_add_param}
\alias{train_add_param}
\title{get (ML) model param}
\usage{
train_add_param(trainObj)
}
\arguments{
\item{train_obj}{caret::train obj}
}
\value{
str
}
\description{
get (ML) model param
}
