% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{train_obsVSpred_plot}
\alias{train_obsVSpred_plot}
\title{scatterplot of obs vs pred from modified version of caret::train
(with items: add_trainDf, add_testDf, add_gof)}
\usage{
train_obsVSpred_plot(train_obj)
}
\arguments{
\item{train_obj}{train obj caret::train obj
(updated with add_trainDf, add_testDf, add_gof)}
}
\value{
p(ggplot2) or list of p
}
\description{
scatterplot of obs vs pred from modified version of caret::train
(with items: add_trainDf, add_testDf, add_gof)
}
