% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{train_plot}
\alias{train_plot}
\title{self version of plot.train}
\usage{
train_plot(obj)
}
\arguments{
\item{obj}{caret::train obj.
\itemize{
\item svmRadialSigma
\item rf
\item keras
\item gassprRadial
\item pls
\item SMLR(stepwise linear regression)
}}
}
\value{
p(ggplot2)
}
\description{
self version of plot.train
}
