% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{train_residual_plot}
\alias{train_residual_plot}
\title{Residual plot(facet by along ID and along obs value)}
\usage{
train_residual_plot(train_obj)
}
\arguments{
\item{train_obj}{train obj(updated with add_trainDf, add_testDf, add_gof)}
}
\value{
p(ggplot2)
}
\description{
Residual plot(facet by along ID and along obs value)
}
