% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{train_update}
\alias{train_update}
\title{update train obj (add_trainDf, add_testDf, add_gof)}
\usage{
train_update(train_obj, spc_inTrain, spc_Test, newdata, biochemphy,
  newdata_inTrain = NULL)
}
\arguments{
\item{train_obj}{caret::train obj}

\item{spc_inTrain}{Speclib obj where Calibration dataset from}

\item{spc_Test}{Speclib obj where Validation dataset from}

\item{newdata}{in predict(fit, newdata = newdata)}

\item{biochemphy}{y name in SI}

\item{newdata_inTrain}{in predict(fit, newdata = newdata)}
}
\value{
updated caret::train Obj
}
\description{
update train obj (add_trainDf, add_testDf, add_gof)
}
