% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{trains_gof_dotplot}
\alias{trains_gof_dotplot}
\title{dotplot of fits gof df}
\usage{
trains_gof_dotplot(input, tagName = "method")
}
\arguments{
\item{input}{multirow gof df}

\item{tagName}{ID col name(group tag)}
}
\value{
p(ggplot2)
}
\description{
dotplot of fits gof df
}
