% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{trains_obsVSPred_plot}
\alias{trains_obsVSPred_plot}
\title{wrapper of \code{\link{train_obsVSpred_plot}} for list of caret::train objs}
\usage{
trains_obsVSPred_plot(train_objs, ...)
}
\arguments{
\item{train_objs}{list of caret::train obj
(updated with add_trainDf, add_testDf, add_gof)}
}
\value{
p(ggplot) or list of p
}
\description{
wrapper of \code{\link{train_obsVSpred_plot}} for list of caret::train objs
}
