% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4plot.R
\name{trains_plot}
\alias{trains_plot}
\title{wrapper of train_plot for list of train objs}
\usage{
trains_plot(objs)
}
\arguments{
\item{objs}{list of train objs}
}
\value{
list of p(ggplot)
}
\description{
wrapper of train_plot for list of train objs
}
