% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4tidymodels.R
\name{workflow_curve_plot}
\alias{workflow_curve_plot}
\title{plot the fitted curve}
\usage{
workflow_curve_plot(fitResult)
}
\arguments{
\item{fitResult}{tidymodel obj}
}
\value{
a ggplot obj
}
\description{
plot the fitted curve
}
