% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4tidymodels.R
\name{workflow_post_plot}
\alias{workflow_post_plot}
\title{obs vs pred scatter plot}
\usage{
workflow_post_plot(fitResult)
}
\arguments{
\item{fitResult}{the tidymodel obj}
}
\value{
a ggplot obj
}
\description{
obs vs pred scatter plot
}
