% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4caret.R
\name{doCaret}
\alias{doCaret}
\title{caret::train wrapper}
\usage{
doCaret(method, tuneGrid, tuneLength, fm, df_inTrain, df_Test, biochemphyNM)
}
\arguments{
\item{method}{model(list) or model name}

\item{tuneGrid}{caret tuneGrid (df)}

\item{tuneLength}{caret tuneLength (numeric)}

\item{fm}{formula}

\item{biochemphyNM}{y name for label}

\item{df}{data}
}
\value{
modified caret::train obj
}
\description{
caret::train wrapper
}
