% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_2dfB}
\alias{spc_2dfB}
\title{Speclib obj to wide df (reflect col with 'B_' prefix)}
\usage{
spc_2dfB(spc)
}
\arguments{
\item{spc}{the Speclib obj}
}
\value{
df (wide df)
}
\description{
Speclib obj to wide df (reflect col with 'B_' prefix)
}
