% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulits4spc.R
\name{spc_resample}
\alias{spc_resample}
\title{resample a spc to same spectral resolution and bands with a given spc}
\usage{
spc_resample(sim, obs)
}
\arguments{
\item{sim}{the spc to be resampled}

\item{obs}{the example spc}
}
\value{
the resampled spc
}
\description{
resample a spc to same spectral resolution and bands with a given spc
}
