% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVI.R
\name{ssdxj_vegindex}
\alias{ssdxj_vegindex}
\title{vegetion index calculation handler}
\usage{
ssdxj_vegindex(index, spc, weighted = FALSE, ...)
}
\arguments{
\item{index:}{name of index, used to match function}

\item{spc:}{the speclib obj}
}
\value{
the index value (vector)
}
\description{
vegetion index calculation handler
}
